backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
motor==3.3.2
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
httpx==0.25.2
pillow==10.1.0
python-jose[cryptography]==3.3.0